Tests: add world load benchmark (#2768)
